requirements.txt
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
tomli==2.0.1
Werkzeug==3.0.1
